+++ cmp/grid/columns/Column.js
@@ -188,8 +188,6 @@ export class Column {
     setValueFn;
     /** @member {Column~getValueFn} */
     getValueFn;
-    /** @member {boolean} */
-    hasCustomGetValueFn;
 
     /** @member {GridModel} */
     gridModel;
@@ -498,7 +496,6 @@ export class Column {
         this.editorIsPopup = editorIsPopup;
         this.setValueFn = withDefault(setValueFn, this.defaultSetValueFn);
         this.getValueFn = withDefault(getValueFn, this.defaultGetValueFn);
-        this.hasCustomGetValueFn = !!getValueFn;
 
         this.gridModel = gridModel;
         this.agOptions = agOptions ? clone(agOptions) : {};