+++ svc/GridExportService.js
@@ -25,12 +25,12 @@ import {
     sortBy,
     compact,
     findIndex,
-    keys, isBoolean, isDate, isInteger, isNumber, isObject
+    keys, isBoolean, isDate, isNumber
 } from 'lodash';
 import {span, a} from '@xh/hoist/cmp/layout';
 import {wait} from '@xh/hoist/promise';
 
-const {AUTO, DATE, LOCAL_DATE} = FieldType;
+const {AUTO, BOOL, DATE, INT, LOCAL_DATE, NUMBER, STRING, PWD} = FieldType;
 
 /**
  * Exports Grid data to either Excel or CSV via Hoist's server-side export capabilities.
@@ -162,8 +162,9 @@ export class GridExportService extends HoistService {
      * @return {string|Object} - value suitable for export to excel, csv, or clipboard.
      */
     getExportableValueForCell({gridModel, record, column, node, forExcel = false}) {
-        const {field, exportValue, getValueFn, hasCustomGetValueFn} = column,
-            aggData = node && gridModel.treeMode && !isEmpty(record.children) ? node.aggData : null;
+        const {field, exportValue, getValueFn, defaultGetValueFn} = column,
+            aggData = node && gridModel.treeMode && !isEmpty(record.children) ? node.aggData : null,
+            hasCustomGetValueFn = getValueFn !== defaultGetValueFn;
 
         // 0) Main processing
         let value = getValueFn({record, field, column, gridModel});
@@ -192,13 +193,15 @@ export class GridExportService extends HoistService {
         const exportFieldType = this.getExportFieldType(column);
 
         let cellSpecificType = null;
-        if (exportFieldType === FieldType.AUTO || isFunction(exportValue) || hasCustomGetValueFn) {
-            cellSpecificType = this.getTypeIfDifferent(value, exportFieldType);
+        if (exportFieldType === AUTO || isFunction(exportValue) || hasCustomGetValueFn) {
+            cellSpecificType = this.getCellSpecificType(value, exportFieldType);
         }
 
         // 2) Dates: Provide the date data string expected by the server endpoint
         // Also functions as a consistent human-friendly date format for CSV and clipboard
-        if (exportFieldType === DATE) value = fmtDate(value, 'YYYY-MM-DD HH:mm:ss');
+        if (exportFieldType === DATE || cellSpecificType === DATE) {
+            value = fmtDate(value, 'YYYY-MM-DD HH:mm:ss');
+        }
 
         value = value.toString();
 
@@ -393,18 +396,14 @@ export class GridExportService extends HoistService {
     }
 
     // Return a value's data type if different from the type specified
-    getTypeIfDifferent(v, type) {
-        const ifTypeNot = (allowedTypes, retType) => allowedTypes.includes(type) ? null : retType,
-            {BOOL, INT, NUMBER, LOCAL_DATE, DATE, STRING, PWD, AUTO, TAGS, JSON} = FieldType;
+    getCellSpecificType(v, colType) {
+        const ifTypeNot = (allowedTypes, retType) => allowedTypes.includes(colType) ? null : retType;
 
         if (isBoolean(v))   return ifTypeNot([BOOL], BOOL);
-        if (isInteger(v))   return ifTypeNot([NUMBER, INT], INT);
-        if (isNumber(v))    return ifTypeNot([NUMBER], NUMBER);
+        if (isNumber(v))    return ifTypeNot([NUMBER, INT], NUMBER);
         if (isLocalDate(v)) return ifTypeNot([LOCAL_DATE], LOCAL_DATE);
         if (isDate(v))      return ifTypeNot([DATE], DATE);
         if (isString(v))    return ifTypeNot([PWD, STRING, AUTO], AUTO);
-        if (isArray(v))     return ifTypeNot([TAGS, JSON, AUTO], AUTO);
-        if (isObject(v))    return ifTypeNot([JSON, AUTO], AUTO);
 
         return null;
     }